% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscquery.R
\name{dscquery}
\alias{dscquery}
\title{SQL-like interface for querying DSC output.}
\usage{
dscquery(dsc.outdir, targets, conditions = NULL, groups, add.path = FALSE,
  exec = "dsc-query", max.extract.vector = 10, verbose = TRUE)
}
\arguments{
\item{dsc.outdir}{Directory where DSC output is stored.}

\item{targets}{Query targets specified as a character vector, e.g.,
\code{targets = c("simulate.n","estimate","mse.score")}. These will
be the names of the columns in the returned data frame.}

\item{conditions}{The default \code{NULL} means "no conditions", in
which case the results for all DSC pipelines are returned.}

\item{groups}{Definition of module groups. This feature is not yet
implemented.}

\item{add.path}{If TRUE, the returned data frame will contain full
pathnames, not just the base filenames.}

\item{exec}{The command or pathname of the dsc-query executable.}

\item{max.extract.vector}{All vectors not exceeding this length are
automatically extracted to the data frame unless the}

\item{verbose}{If \code{verbose = TRUE}, print progress of DSC
query command to the console.}
}
\value{
A data frame containing the result of the DSC query, with
columns corresponding to the query target.

NOTE: It is possible that some of the column names may be
duplicated if any vectors are extracted.
}
\description{
Add more detailed (paragraph-length) description here.
}
\note{
May not work in Windows.
}
\examples{

# Retrieve results from the "one_sample_location" experiment in
# which the true mean is 1. The MSE (mean squared error) values
# should be extracted into the "mse.mse" column.
dsc.dir <- system.file("datafiles","one_sample_location",
                       "dsc_result",package = "dscrutils")
dat <- dscquery(dsc.dir,targets = c("simulate.n","estimate","mse.mse"),
                condition = "simulate.true_mean = 1")
print(dat)

# This query should generate an error because there is no output
# called "score" in the "mse" module.
dat2 <- dscquery(dsc.dir,targets = c("simulate.n","estimate","mse.score"),
                 condition = "simulate.true_mean = 1")

}
