% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscquery.R
\name{dscquery}
\alias{dscquery}
\title{dsc.dir <- system.file("datafiles","one_sample_location",
                       "dsc_result",package = "dscrutils")
dat1 <- dscquery(dsc.dir,
                 targets = c("simulate.n","analyze","score.error"))}
\usage{
dscquery(dsc.outdir, targets = NULL, targets.notreq = NULL,
  conditions = NULL, groups = NULL, return.type = c("auto",
  "data.frame", "list"), ignore.missing.files = FALSE,
  omit.filenames = TRUE, exec = "dsc-query", verbose = TRUE)
}
\description{
# Retrieve the results only for simulations in which the "mean" module
# was run. Because this is a condition for a module name, it is
# applied before loading the full set of results into R. Therefore,
# this type of filtering step can speed up the query when there are
# many simulation results.
dat2 <- dscquery(dsc.dir,
                 targets = c("simulate.n","analyze","score.error"),
                 conditions = "$(analyze) == 'mean'")
}
\details{
# Return results only for simulations in which the error summary is
# greater than 0.2. This condition is applied only after loading the
# full set of results into R. Therefore, this type of condition will not
# reduce the query runtime.
dat3 <- dscquery(dsc.dir,
                 targets = c("simulate.n","analyze","score.error"),
                 conditions = "$(score.error) > 0.2")

# Retrieve the DSC results only for simulations in which the "mean"
# module was run, and which which the error summary is greater than
# 0.2. The conditions in this case are applied both before and after
# loading results into R.
dat4 <- dscquery(dsc.dir,
                 targets = c("simulate.n","analyze","score.error"),
                 conditions = c("$(score.error) > 0.2",
                                "$(analyze) == 'median'"))

# Retrieve some results from the "ash" DSC experiment. In this
# example, the beta estimates are vectors, so the results are
# extracted into a list by default.
dsc.dir2 <- system.file("datafiles","ash","dsc_result",
                        package = "dscrutils")
dat5 <-
  dscquery(dsc.dir2,
           targets = c("simulate.nsamp","simulate.g","shrink.mixcompdist",
                       "shrink.beta_est","shrink.pi0_est"),
           conditions = "$(simulate.g)=='list(c(2/3,1/3),c(0,0),c(1,2))'")

# This is the same as the previous example, but extracts the results
# into data frame. Since the vectors cannot be stored in a data frame,
# the names of the files storing the vectors are returned instead.
dat6 <-
  dscquery(dsc.dir2,
           targets = c("simulate.nsamp","simulate.g","shrink.mixcompdist",
                       "shrink.beta_est","shrink.pi0_est"),
           conditions = "$(simulate.g)=='list(c(2/3,1/3),c(0,0),c(1,2))'",
           return.type = "data.frame")
}
