% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscquery.R
\name{dscquery}
\alias{dscquery}
\title{SQL-like interface for querying DSC output.}
\usage{
dscquery(dsc.outdir, targets, conditions = NULL, groups, add.path = FALSE,
  exec = "dsc-query", verbose = TRUE)
}
\arguments{
\item{dsc.outdir}{Directory where DSC output is stored.}

\item{targets}{Query targets specified as a character vector, e.g.,
\code{targets = c("simulate.n","estimate","mse.score")}. These will
be the names of the columns in the returned data frame.}

\item{conditions}{The default \code{NULL} means "no conditions", in
which case the results for all DSC pipelines are returned.}

\item{groups}{Definition of module groups. Currently}

\item{add.path}{If TRUE, the returned data frame will contain full
pathnames, not just the base filenames.}

\item{exec}{The command or pathname of the dsc-query executable.}

\item{verbose}{If \code{verbose = TRUE}, print progress of DSC
query command to the console.}
}
\value{
A data frame containing the result of the DSC query, with
columns corresponding to the query target.
}
\description{
Add more detailed (paragraph-length) description here.
}
\note{
May not work in Windows.
}
\examples{
// Add an example here.

}
