% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscread.R
\name{dscread}
\alias{dscread}
\title{Add title here.}
\usage{
dscread(outdir, outfile)
}
\arguments{
\item{outdir}{Directory where the DSC output is stored.}

\item{outfile}{Describe argument "outfile" here.}
}
\value{
Describe return value here.
}
\description{
Add description here.
}
\examples{

# Add one or more examples building on 

}
\seealso{
\code{\link{dscquery}}
}
