% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscreval.R
\name{dscreval}
\alias{dscreval}
\title{Parse and evaluate R expression, and check that the value is
a "simple" atomic object, or a list of simple atomic objects.}
\usage{
dscreval(x)
}
\description{
Parse and evaluate R expression, and check that the value is
a "simple" atomic object, or a list of simple atomic objects.
}
\examples{

# Generates a numeric vector.
dscreval("c(3,-1,14)")

# Generates a numeric vector.
dscreval("seq(1,2,length.out = 5)")

# Generates a logical vector.
dscreval("1:10 < 5")

# Generates a list with two vectors.
dscreval("list(x = LETTERS[1:5],y = 1:5)")

\dontrun{
# Generates an error.
dscreval("NULL")

# Generates an error.
dscreval("polyroot(c(-1,2,-1,4))")

# Generates an error.
dscreval("vector('list',3)")

# Generates an error.
dscreval("list(x = LETTERS[1:5],y = 1:5,z = as.list(1:5))")
}

}
